Rename UnittestDataTree to DataTree and move to own file
